move pluginLogsLogic to pipeline folder
