+++ frontend/src/scenes/pipeline/PipelineAppLogs.tsx
@@ -4,11 +4,7 @@ import { LOGS_PORTION_LIMIT } from 'lib/constants'
 import { dayjs } from 'lib/dayjs'
 import { pluralize } from 'lib/utils'
 
-import {
-    PipelineAppLogLevel,
-    pluginLogsLogic,
-    PluginLogsProps as PipelineAppLogsProps,
-} from '../plugins/plugin/pluginLogsLogic'
+import { PipelineAppLogLevel, pluginLogsLogic, PluginLogsProps as PipelineAppLogsProps } from './pipelineAppLogsLogic'
 
 function LogLevelDisplay(type: PipelineAppLogLevel): JSX.Element {
     let color: string | undefined
