+++ frontend/src/scenes/pipeline/pipelineAppLogsLogic.ts
